requirements.txt
pyrofork
tgcrypto
python-dotenv
aiohttp
requests
pytube
aiofiles
pymongo
asyncio
beautifulsoup4
dnspython
ffmpeg-python
gitpython
hachoir
heroku3
motor
pillow
psutil
pyfiglet
py-tgcalls==0.9.7
pykeyboard
pyyaml
httpx[http2]
speedtest-cli
spotipy
unidecode
yt-dlp
youtube-search
youtube-search-python
tgcrypto
flask
flask_restful
gunicorn